requirements.txt
annotated-types==0.7.0
anyio==4.5.2
apache-airflow==2.9.1
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
autopep8==2.3.1
backcall==0.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
dacite==1.9.2
debugpy==1.8.14
decorator==5.2.1
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fonttools==4.57.0
h11==0.16.0
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
multimethod==1.10
nest-asyncio==1.6.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
packaging==25.0
pandas>=2.0.0
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
phik==0.12.4
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil>=5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger>=2.0.4
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.4.0
requests==2.32.3
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.44.0
statsmodels==0.14.1
tomli>=2.0.0
tornado==6.4.2
tqdm>=4.67.0
traitlets==5.14.3
typeguard==4.4.0
typer>=0.9.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
visions==0.8.1
wcwidth==0.2.13
wordcloud==1.9.4
ydata-profiling==4.16.1
zipp==3.20.2
